notary_agent/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pdfkit
jinja2
